backend-payload/scripts/requirements.txt
# Python dependencies for image processing scripts
rembg[cpu]>=2.0.50
pillow>=10.0.0
